requirements.txt
groq
numpy
python-dotenv
sounddevice
soundfile
openai-whisper
whisperlivekit
pyannote.audio
torch
torchaudio